Refactor the PyG TensorNet components to ensure compatibility with pure PyTorch.
