+++ tests/graph/test_compute_pyg.py
@@ -60,7 +60,7 @@ class TestCompute:
         lattice = torch.tensor(s1.lattice.matrix, dtype=matgl.float_th).unsqueeze(dim=0)
         g1.pbc_offshift = torch.matmul(g1.pbc_offset, lattice[0])
         g1.pos = g1.frac_coords @ lattice[0]
-        bv, _ = compute_pair_vector_and_distance(g1)
+        bv, _ = compute_pair_vector_and_distance(g1.pos, g1.edge_index, g1.pbc_offshift)
         g1.bond_vec = bv
         d = torch.linalg.norm(g1.bond_vec, axis=1)
 
@@ -73,7 +73,7 @@ class TestCompute:
         lattice = torch.tensor(np.identity(3), dtype=matgl.float_th).unsqueeze(dim=0)
         g2.pbc_offshift = torch.matmul(g2.pbc_offset, lattice[0])
         g2.pos = g2.frac_coords @ lattice[0]
-        bv, _ = compute_pair_vector_and_distance(g2)
+        bv, _ = compute_pair_vector_and_distance(g2.pos, g2.edge_index, g2.pbc_offshift)
         g2.bond_vec = bv
         d = torch.linalg.norm(g2.bond_vec, axis=1)
